+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- The `inf` histogram bucket label will now be exported as `+Inf` in line with the other clients
+
 ## 0.0.8
 
 - `Histogram` now passes custom buckets to child instances